+++ Sources/Scenes/Search/SearchPresenter.swift
@@ -2,7 +2,7 @@ import Foundation
 
 protocol SearchPresenterOutput {
     func display(users: [APIClient.User])
-    func display(nextPage: [APIClient.User])
+    func display(groups: [APIClient.Group])
     func displaySearchError()
 }
 
@@ -15,12 +15,12 @@ class SearchPresenter {
 }
 
 extension SearchPresenter: SearchInteractorOutput {
-    func didFetch(users: [APIClient.User]) {
-        output.display(users: users)
+    func didFetch(groups: [APIClient.Group]) {
+        output.display(groups: groups)
     }
 
-    func didFetch(nextPage: [APIClient.User]) {
-        output.display(nextPage: nextPage)
+    func didFetch(users: [APIClient.User]) {
+        output.display(users: users)
     }
 
     func failedToFetch() {